+++ components/gfx/platform/macos/font_context.rs
@@ -2,7 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#[derive(Clone, Debug, MallocSizeOf)]
+use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
+
+#[derive(Clone, Debug)]
 pub struct FontContextHandle {
     ctx: ()
 }
@@ -11,5 +13,11 @@ impl FontContextHandle {
     // this is a placeholder until NSFontManager or whatever is bound in here.
     pub fn new() -> FontContextHandle {
         FontContextHandle { ctx: () }
+    }
+}
+
+impl MallocSizeOf for FontContextHandle {
+    fn size_of(&self, _: &mut MallocSizeOfOps) -> usize {
+        0
     }
 }
