+++ components/layout/inline.rs
@@ -10,6 +10,7 @@ use block::AbsoluteAssignBSizesTraversal;
 use context::{LayoutContext, LayoutFontContext};
 use display_list::{DisplayListBuildState, InlineFlowDisplayListBuilding};
 use display_list::StackingContextCollectionState;
+use display_list::items::OpaqueNode;
 use euclid::{Point2D, Size2D};
 use floats::{FloatKind, Floats, PlacementInfo};
 use flow::{BaseFlow, Flow, FlowClass, ForceNonfloatedFlag};
@@ -18,7 +19,6 @@ use flow_ref::FlowRef;
 use fragment::{CoordinateSystem, Fragment, FragmentBorderBoxIterator, Overflow};
 use fragment::FragmentFlags;
 use fragment::SpecificFragmentInfo;
-use gfx::display_list::OpaqueNode;
 use gfx::font::FontMetrics;
 use gfx_traits::print_tree::PrintTree;
 use layout_debug;
