+++ components/layout/opaque_node.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use gfx::display_list::OpaqueNode;
+use display_list::items::OpaqueNode;
 use libc::c_void;
 use script_traits::UntrustedNodeAddress;
