+++ components/layout/query.rs
@@ -8,10 +8,10 @@ use app_units::Au;
 use construct::ConstructionResult;
 use context::LayoutContext;
 use display_list::IndexableText;
+use display_list::items::{DisplayList, OpaqueNode, ScrollOffsetMap};
 use euclid::{Point2D, Vector2D, Rect, Size2D};
 use flow::{Flow, GetBaseFlow};
 use fragment::{Fragment, FragmentBorderBoxIterator, SpecificFragmentInfo};
-use gfx::display_list::{DisplayList, OpaqueNode, ScrollOffsetMap};
 use inline::InlineFragmentNodeFlags;
 use ipc_channel::ipc::IpcSender;
 use msg::constellation_msg::PipelineId;
