+++ tests/unit/metrics/lib.rs
@@ -4,12 +4,11 @@
 
 #![cfg(test)]
 
-extern crate gfx;
 extern crate gfx_traits;
 extern crate ipc_channel;
+extern crate layout;
 extern crate metrics;
 extern crate msg;
-extern crate net_traits;
 extern crate profile_traits;
 extern crate servo_url;
 extern crate time;
