+++ tests/unit/metrics/paint_time.rs
@@ -2,17 +2,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use gfx::display_list::{BaseDisplayItem, WebRenderImageInfo};
-use gfx::display_list::{DisplayItem, DisplayList, ImageDisplayItem};
 use gfx_traits::Epoch;
 use ipc_channel::ipc;
+use layout::display_list::items::{BaseDisplayItem, DisplayItem, DisplayList, ImageDisplayItem};
 use metrics::{PaintTimeMetrics, ProfilerMetadataFactory, ProgressiveWebMetric};
 use msg::constellation_msg::TEST_PIPELINE_ID;
-use net_traits::image::base::PixelFormat;
 use profile_traits::time::{ProfilerChan, TimerMetadata};
 use servo_url::ServoUrl;
 use time;
-use webrender_api::{ImageRendering, LayoutSize};
+use webrender_api::{ImageKey, ImageRendering, LayoutSize};
 
 struct DummyProfilerMetadataFactory {}
 impl ProfilerMetadataFactory for DummyProfilerMetadataFactory {
@@ -120,12 +118,7 @@ fn test_first_paint_setter() {
 fn test_first_contentful_paint_setter() {
     let image = DisplayItem::Image(Box::new(ImageDisplayItem {
         base: BaseDisplayItem::empty(),
-        webrender_image: WebRenderImageInfo {
-            width: 1,
-            height: 1,
-            format: PixelFormat::RGB8,
-            key: None,
-        },
+        id: ImageKey::DUMMY,
         stretch_size: LayoutSize::zero(),
         tile_spacing: LayoutSize::zero(),
         image_rendering: ImageRendering::Auto,
