+++ tests/wpt/web-platform-tests/css/css-backgrounds/reference/background-rounded-image-clip.html
@@ -1,0 +1,15 @@
+<!doctype html>
+<style>
+    html {
+        background-color: green;
+    }
+    #a {
+        top: 20px;
+        left: 20px;
+        position: absolute;
+        width: 20px;
+        height: 20px;
+        background-color: black;
+    }
+</style>
+<div id="a"></div>
